+++ src/Ant3Arena.Business/Ants/AntBlack.cs
@@ -2,23 +2,22 @@
 using System;
 using System.Drawing;
 
-namespace Ant_3_Arena.Ants
+namespace Ant3Arena.Business.Ants
 {
-	public class AntYellow : IAnt
+	public class AntBlack : IAnt
 	{
 		public int X { get; set; }
 		public int Y { get; set; }
 		private string Direction { get; set; }
-		private int Verticalvelocity = 4;
-		private int Horizontalvelocity = 4;
-		private readonly string color = "#FFFF00";
+		private int Verticalvelocity = 6;
+		private int Horizontalvelocity = 2;
+		private readonly string color = "#000000";
 		private readonly Bitmap antImage;
-
 		public Bitmap AntImage { get { return antImage; } }
 
-		public AntYellow(Size borders)
+		public AntBlack(Size borders)
 		{
-			Direction = "RightDown";
+			Direction = "RightUp";
 			Color newColor = ColorTranslator.FromHtml(color);
 			Color white = ColorTranslator.FromHtml("#FFFFFF");
 
@@ -41,18 +40,19 @@ namespace Ant_3_Arena.Ants
 			Y = random.Next(0, borders.Height);
 		}
 
-		public void Move(Size borders){
+		public void Move(Size borders)
+		{
 			switch (Direction)
 			{
 				case "LeftUp":
 					X = X - Horizontalvelocity;
 					Y = Y - Verticalvelocity;
 
-					if (X < 0 && Y < 0)
+					if (X < 0 && Y < 500)
 						Direction = "RightDown";
 					else if (X < 0)
 						Direction = "RightUp";
-					else if (Y < 0)
+					else if (Y < 500)
 						Direction = "LeftDown";
 					break;
 				case "LeftDown":
@@ -70,11 +70,11 @@ namespace Ant_3_Arena.Ants
 					X = X + Horizontalvelocity;
 					Y = Y - Verticalvelocity;
 
-					if (X > borders.Width && Y < 0)
+					if (X > borders.Width && Y < 500)
 						Direction = "LeftDown";
 					else if (X > borders.Width)
 						Direction = "LeftUp";
-					else if (Y < 0)
+					else if (Y < 500)
 						Direction = "RightDown";
 					break;
 				case "RightDown":
