+++ src/Ant3Arena.Business/Resources/Ant.png
