+++ src/Tests/Ants3Arena.Business.UnitTests/Ants/AntRed.Tests.cs
@@ -1,4 +1,4 @@
-﻿using Ant_3_Arena.Ants;
+﻿using Ant3Arena.Business.Ants;
 using System.Drawing;
 using Xunit;
 
@@ -18,6 +18,30 @@ namespace Ants3Arena.UnitTests.Ants
             Assert.InRange(antRed.Y, 0, borders.Height);
             Assert.NotNull(antRed.AntImage);
 		}
+
+        [Fact]
+        public static void AntBlack_Should_Be_Black()
+        {
+            // Arrange
+            var borders = new Size(800, 450);
+
+            // Act
+            var antBlack = new AntBlack(borders);
+
+            // Assert
+            for (int x = 0; x < antBlack.AntImage.Width; x++)
+            {
+                for (int y = 0; y < antBlack.AntImage.Height; y++)
+                {
+                    Color pixelColor = antBlack.AntImage.GetPixel(x, y);
+                    if (pixelColor.A == 0 || pixelColor.R == 0)
+                    {
+                        continue; // Skip transparent and black pixels (black is the outline)
+                    }
+                    Assert.True(pixelColor.ToString() == ColorTranslator.FromHtml("#FF0000").ToString(), $"Not all white pixels are converted to Red. X,Y: {x},{y} has color: {pixelColor.ToString()}.");
+                }
+            }
+        }
 
         [Fact]
         public static void AntRed_Move_Works()
