+++ src/Tests/Ants3Arena.Business.UnitTests/Ants3Arena.Business.UnitTests.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net9.0-windows</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <IsPackable>false</IsPackable>
@@ -15,6 +15,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\Ants3Arena.FrontEnd\Ant 3 Arena.csproj" />
+    <ProjectReference Include="..\..\Ant3Arena.Business\Ant3Arena.Business.csproj" />
   </ItemGroup>
 </Project>
